requirements.txt
asgiref==3.9.1
boto3==1.34.69
botocore==1.34.162
Django==5.2.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.2
django-filter>=23.5
django-storages==1.14.2
drf-yasg>=1.21.7
gunicorn==21.2.0
jmespath==1.0.1
Markdown>=3.5
mysqlclient==2.2.7
pillow==10.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML>=6.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate>=4.1
urllib3==2.5.0